# server/docs/swagger.yaml
definitions:
...
  handler.Error:
    properties:
      code:
        type: integer
      error:
        type: string
      errors:
        items:
          type: string
        type: array
      input_errors:
        additionalProperties:
          type: string
        type: object
    type: object
  request.ChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  request.ChangeUsername:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  request.ConfirmEmail:
    properties:
      code:
        type: string
    type: object
...
  request.DeleteUser:
    properties:
      password:
        type: string
    type: object
  request.EmailChangeRequest:
    properties:
      email:
        type: string
    type: object
...
  request.UserSignIn:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  request.UserSignUp:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  response.Status:
    properties:
      success:
        type: boolean
    type: object
  response.UserSignIn:
    properties:
      id:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users/:
    delete:
      consumes:
      - application/json
      operationId: DeleteUser
      parameters:
      - description: Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.DeleteUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "422":
          description: Validation error or incorrect password
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete user account
      tags:
      - users
  /users/confirm-email/:
    post:
      consumes:
      - application/json
      operationId: ConfirmEmail
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ConfirmEmail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Confirm email
      tags:
      - users
  /users/email/:
    post:
      consumes:
      - application/json
      operationId: EmailChangeConfirm
      parameters:
      - description: Old and new passwords
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.EmailChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Confirm changing user email
      tags:
      - users
...
  /users/password/:
    post:
      consumes:
      - application/json
      operationId: ChangePassword
      parameters:
      - description: Old and new passwords
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "422":
          description: Validation error or incorrect old password
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - users
  /users/sign-in/:
    post:
      consumes:
      - application/json
      operationId: UserSignIn
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UserSignIn'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: User auth
      tags:
      - users
  /users/sign-out/:
    post:
      operationId: UserSignOut
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - users
  /users/sign-up/:
    post:
      consumes:
      - application/json
      operationId: UserSignUp
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: User registration
      tags:
      - users
  /users/username/:
    post:
      consumes:
      - application/json
      operationId: ChangeUsername
      parameters:
      - description: New username and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ChangeUsername'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Status'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.Error'
        "422":
          description: Validation error or incorrect password
          schema:
            $ref: '#/definitions/handler.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Error'
      security:
      - ApiKeyAuth: []
      summary: Change username
      tags:
      - users
swagger: "2.0"